backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
django-filter==23.3
drf-yasg==1.21.7
PyJWT==2.8.0
setuptools==69.0.0
